requirements.txt
flask
openai
python-dotenv
rapidfuzz
paho-mqtt
mysql-connector